style: api 폴더 분리
