+++ src/api/instance.ts
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+const instance = axios.create({
+  baseURL: "/api",
+});
+
+export default instance;
