+++ src/api/question.ts
@@ -1,0 +1,10 @@
+import { Question } from "../@types";
+import instance from "./instance";
+
+export const getQuestions = (params?: { date?: string }) => {
+  return instance.get<Question[]>("/api/questions", { params });
+};
+
+export const getQuestion = (id: string) => {
+  return instance.get<Question>("/api/questions/" + id);
+};
